third_party/WebKit/LayoutTests/platform/win/virtual/web-animations-css/animations/transition-and-animation-2-expected.txt

Warning this test is running in real-time and may be flaky.
FAIL - "webkitTransform.5" property for "box" element at 0.4s expected: 0 but saw: matrix(1, 0, 0, 1, 400, 100)

Log:
9.000000000014552 - runAnimationTest
10.999999999967258 - First webkitAnimationStart event fired
10.999999999967258 - Finished waiting for animations to start
10.999999999967258 - RAF callback, animation time: 0
10.999999999967258 - RAF callback, animation time: 0
11.999999999943611 - RAF callback, animation time: 0.9999999999763531
12.999999999919964 - RAF callback, animation time: 1.9999999999527063
13.999999999896318 - RAF callback, animation time: 2.9999999999290594
15.000000000100044 - RAF callback, animation time: 4.000000000132786
16.000000000076398 - RAF callback, animation time: 5.000000000109139
17.00000000005275 - RAF callback, animation time: 6.0000000000854925
18.000000000029104 - RAF callback, animation time: 7.000000000061846
19.000000000005457 - RAF callback, animation time: 8.000000000038199
19.99999999998181 - RAF callback, animation time: 9.000000000014552
20.999999999958163 - RAF callback, animation time: 9.999999999990905
21.999999999934516 - RAF callback, animation time: 10.999999999967258
22.99999999991087 - RAF callback, animation time: 11.999999999943611
23.999999999887223 - RAF callback, animation time: 12.999999999919964
25.00000000009095 - RAF callback, animation time: 14.000000000123691
26.000000000067303 - RAF callback, animation time: 15.000000000100044
27.000000000043656 - RAF callback, animation time: 16.000000000076398
28.00000000002001 - RAF callback, animation time: 17.00000000005275
28.999999999996362 - RAF callback, animation time: 18.000000000029104
29.999999999972715 - RAF callback, animation time: 19.000000000005457
30.99999999994907 - RAF callback, animation time: 19.99999999998181
31.99999999992542 - RAF callback, animation time: 20.999999999958163
32.999999999901775 - RAF callback, animation time: 21.999999999934516
34.0000000001055 - RAF callback, animation time: 23.000000000138243
35.000000000081855 - RAF callback, animation time: 24.000000000114596
36.00000000005821 - RAF callback, animation time: 25.00000000009095
37.00000000003456 - RAF callback, animation time: 26.000000000067303
38.000000000010914 - RAF callback, animation time: 27.000000000043656
38.99999999998727 - RAF callback, animation time: 28.00000000002001
39.99999999996362 - RAF callback, animation time: 28.999999999996362
40.99999999993997 - RAF callback, animation time: 29.999999999972715
41.99999999991633 - RAF callback, animation time: 30.99999999994907
42.99999999989268 - RAF callback, animation time: 31.99999999992542
44.000000000096406 - RAF callback, animation time: 33.00000000012915
45.00000000007276 - RAF callback, animation time: 34.0000000001055
46.00000000004911 - RAF callback, animation time: 35.000000000081855
47.000000000025466 - RAF callback, animation time: 36.00000000005821
48.00000000000182 - RAF callback, animation time: 37.00000000003456
48.99999999997817 - RAF callback, animation time: 38.000000000010914
49.999999999954525 - RAF callback, animation time: 38.99999999998727
50.99999999993088 - RAF callback, animation time: 39.99999999996362
51.99999999990723 - RAF callback, animation time: 40.99999999993997
52.999999999883585 - RAF callback, animation time: 41.99999999991633
54.00000000008731 - RAF callback, animation time: 43.00000000012005
55.000000000063665 - RAF callback, animation time: 44.000000000096406
56.00000000004002 - RAF callback, animation time: 45.00000000007276
57.00000000001637 - RAF callback, animation time: 46.00000000004911
58.99999999996908 - RAF callback, animation time: 48.00000000000182
59.99999999994543 - RAF callback, animation time: 48.99999999997817
60.99999999992178 - RAF callback, animation time: 49.999999999954525
61.99999999989814 - RAF callback, animation time: 50.99999999993088
63.00000000010186 - RAF callback, animation time: 52.000000000134605
64.00000000007822 - RAF callback, animation time: 53.00000000011096
65.00000000005457 - RAF callback, animation time: 54.00000000008731
66.00000000003092 - RAF callback, animation time: 55.000000000063665
67.00000000000728 - RAF callback, animation time: 56.00000000004002
67.99999999998363 - RAF callback, animation time: 57.00000000001637
68.99999999995998 - RAF callback, animation time: 57.999999999992724
69.99999999993634 - RAF callback, animation time: 58.99999999996908
70.99999999991269 - RAF callback, animation time: 59.99999999994543
71.99999999988904 - RAF callback, animation time: 60.99999999992178
73.00000000009277 - RAF callback, animation time: 62.00000000012551
74.00000000006912 - RAF callback, animation time: 63.00000000010186
75.00000000004547 - RAF callback, animation time: 64.00000000007822
76.00000000002183 - RAF callback, animation time: 65.00000000005457
76.99999999999818 - RAF callback, animation time: 66.00000000003092
77.99999999997453 - RAF callback, animation time: 67.00000000000728
78.99999999995089 - RAF callback, animation time: 67.99999999998363
79.99999999992724 - RAF callback, animation time: 68.99999999995998
80.9999999999036 - RAF callback, animation time: 69.99999999993634
82.00000000010732 - RAF callback, animation time: 71.00000000014006
83.00000000008367 - RAF callback, animation time: 72.00000000011642
84.00000000006003 - RAF callback, animation time: 73.00000000009277
85.00000000003638 - RAF callback, animation time: 74.00000000006912
86.00000000001273 - RAF callback, animation time: 75.00000000004547
86.99999999998909 - RAF callback, animation time: 76.00000000002183
87.99999999996544 - RAF callback, animation time: 76.99999999999818
88.99999999994179 - RAF callback, animation time: 77.99999999997453
89.99999999991815 - RAF callback, animation time: 78.99999999995089
90.9999999998945 - RAF callback, animation time: 79.99999999992724
92.00000000009823 - RAF callback, animation time: 81.00000000013097
93.00000000007458 - RAF callback, animation time: 82.00000000010732
94.00000000005093 - RAF callback, animation time: 83.00000000008367
95.00000000002728 - RAF callback, animation time: 84.00000000006003
96.00000000000364 - RAF callback, animation time: 85.00000000003638
96.99999999997999 - RAF callback, animation time: 86.00000000001273
97.99999999995634 - RAF callback, animation time: 86.99999999998909
98.9999999999327 - RAF callback, animation time: 87.99999999996544
99.99999999990905 - RAF callback, animation time: 88.99999999994179
100.9999999998854 - RAF callback, animation time: 89.99999999991815
102.00000000008913 - RAF callback, animation time: 91.00000000012187
103.00000000006548 - RAF callback, animation time: 92.00000000009823
104.00000000004184 - RAF callback, animation time: 93.00000000007458
105.00000000001819 - RAF callback, animation time: 94.00000000005093
105.99999999999454 - RAF callback, animation time: 95.00000000002728
106.9999999999709 - RAF callback, animation time: 96.00000000000364
107.99999999994725 - RAF callback, animation time: 96.99999999997999
108.9999999999236 - RAF callback, animation time: 97.99999999995634
109.99999999989996 - RAF callback, animation time: 98.9999999999327
111.00000000010368 - RAF callback, animation time: 100.00000000013642
112.00000000008004 - RAF callback, animation time: 101.00000000011278
113.00000000005639 - RAF callback, animation time: 102.00000000008913
114.00000000003274 - RAF callback, animation time: 103.00000000006548
115.0000000000091 - RAF callback, animation time: 104.00000000004184
115.99999999998545 - RAF callback, animation time: 105.00000000001819
116.9999999999618 - RAF callback, animation time: 105.99999999999454
117.99999999993815 - RAF callback, animation time: 106.9999999999709
118.99999999991451 - RAF callback, animation time: 107.99999999994725
119.99999999989086 - RAF callback, animation time: 108.9999999999236
121.00000000009459 - RAF callback, animation time: 110.00000000012733
122.00000000007094 - RAF callback, animation time: 111.00000000010368
123.0000000000473 - RAF callback, animation time: 112.00000000008004
124.00000000002365 - RAF callback, animation time: 113.00000000005639
125 - RAF callback, animation time: 114.00000000003274
125.99999999997635 - RAF callback, animation time: 115.0000000000091
126.9999999999527 - RAF callback, animation time: 115.99999999998545
127.99999999992906 - RAF callback, animation time: 116.9999999999618
128.9999999999054 - RAF callback, animation time: 117.99999999993815
129.99999999988177 - RAF callback, animation time: 118.99999999991451
131.0000000000855 - RAF callback, animation time: 120.00000000011823
132.00000000006185 - RAF callback, animation time: 121.00000000009459
133.0000000000382 - RAF callback, animation time: 122.00000000007094
134.00000000001455 - RAF callback, animation time: 123.0000000000473
134.9999999999909 - RAF callback, animation time: 124.00000000002365
135.99999999996726 - RAF callback, animation time: 125
136.9999999999436 - RAF callback, animation time: 125.99999999997635
137.99999999991996 - RAF callback, animation time: 126.9999999999527
138.99999999989632 - RAF callback, animation time: 127.99999999992906
140.00000000010004 - RAF callback, animation time: 129.0000000001328
141.0000000000764 - RAF callback, animation time: 130.00000000010914
142.00000000005275 - RAF callback, animation time: 131.0000000000855
143.0000000000291 - RAF callback, animation time: 132.00000000006185
144.00000000000546 - RAF callback, animation time: 133.0000000000382
144.9999999999818 - RAF callback, animation time: 134.00000000001455
145.99999999995816 - RAF callback, animation time: 134.9999999999909
146.99999999993452 - RAF callback, animation time: 135.99999999996726
147.99999999991087 - RAF callback, animation time: 136.9999999999436
148.99999999988722 - RAF callback, animation time: 137.99999999991996
150.00000000009095 - RAF callback, animation time: 139.0000000001237
151.0000000000673 - RAF callback, animation time: 140.00000000010004
152.00000000004366 - RAF callback, animation time: 141.0000000000764
153.00000000002 - RAF callback, animation time: 142.00000000005275
153.99999999999636 - RAF callback, animation time: 143.0000000000291
154.99999999997272 - RAF callback, animation time: 144.00000000000546
155.99999999994907 - RAF callback, animation time: 144.9999999999818
156.99999999992542 - RAF callback, animation time: 145.99999999995816
161.0000000000582 - RAF callback, animation time: 150.00000000009095
163.0000000000109 - RAF callback, animation time: 152.00000000004366
163.99999999998727 - RAF callback, animation time: 153.00000000002
164.99999999996362 - RAF callback, animation time: 153.99999999999636
165.99999999993997 - RAF callback, animation time: 154.99999999997272
166.99999999991633 - RAF callback, animation time: 155.99999999994907
167.99999999989268 - RAF callback, animation time: 156.99999999992542
169.0000000000964 - RAF callback, animation time: 158.00000000012915
170.00000000007276 - RAF callback, animation time: 159.0000000001055
171.0000000000491 - RAF callback, animation time: 160.00000000008185
172.00000000002547 - RAF callback, animation time: 161.0000000000582
173.00000000000182 - RAF callback, animation time: 162.00000000003456
173.99999999997817 - RAF callback, animation time: 163.0000000000109
174.99999999995453 - RAF callback, animation time: 163.99999999998727
175.99999999993088 - RAF callback, animation time: 164.99999999996362
176.99999999990723 - RAF callback, animation time: 165.99999999993997
177.99999999988358 - RAF callback, animation time: 166.99999999991633
179.0000000000873 - RAF callback, animation time: 168.00000000012005
180.00000000006366 - RAF callback, animation time: 169.0000000000964
181.00000000004002 - RAF callback, animation time: 170.00000000007276
182.00000000001637 - RAF callback, animation time: 171.0000000000491
182.99999999999272 - RAF callback, animation time: 172.00000000002547
183.99999999996908 - RAF callback, animation time: 173.00000000000182
184.99999999994543 - RAF callback, animation time: 173.99999999997817
185.99999999992178 - RAF callback, animation time: 174.99999999995453
186.99999999989814 - RAF callback, animation time: 175.99999999993088
188.00000000010186 - RAF callback, animation time: 177.0000000001346
189.00000000007822 - RAF callback, animation time: 178.00000000011096
190.00000000005457 - RAF callback, animation time: 179.0000000000873
191.00000000003092 - RAF callback, animation time: 180.00000000006366
192.00000000000728 - RAF callback, animation time: 181.00000000004002
192.99999999998363 - RAF callback, animation time: 182.00000000001637
193.99999999995998 - RAF callback, animation time: 182.99999999999272
194.99999999993634 - RAF callback, animation time: 183.99999999996908
195.9999999999127 - RAF callback, animation time: 184.99999999994543
196.99999999988904 - RAF callback, animation time: 185.99999999992178
198.00000000009277 - RAF callback, animation time: 187.0000000001255
199.00000000006912 - RAF callback, animation time: 188.00000000010186
200.00000000004547 - RAF callback, animation time: 189.00000000007822
201.00000000002183 - RAF callback, animation time: 190.00000000005457
201.99999999999818 - RAF callback, animation time: 191.00000000003092
202.99999999997453 - RAF callback, animation time: 192.00000000000728
203.9999999999509 - RAF callback, animation time: 192.99999999998363
204.99999999992724 - RAF callback, animation time: 193.99999999995998
205.9999999999036 - RAF callback, animation time: 194.99999999993634
207.00000000010732 - RAF callback, animation time: 196.00000000014006
208.00000000008367 - RAF callback, animation time: 197.00000000011642
209.00000000006003 - RAF callback, animation time: 198.00000000009277
210.00000000003638 - RAF callback, animation time: 199.00000000006912
211.00000000001273 - RAF callback, animation time: 200.00000000004547
211.9999999999891 - RAF callback, animation time: 201.00000000002183
212.99999999996544 - RAF callback, animation time: 201.99999999999818
213.9999999999418 - RAF callback, animation time: 202.99999999997453
214.99999999991815 - RAF callback, animation time: 203.9999999999509
215.9999999998945 - RAF callback, animation time: 204.99999999992724
217.00000000009823 - RAF callback, animation time: 206.00000000013097
218.00000000007458 - RAF callback, animation time: 207.00000000010732
219.00000000005093 - RAF callback, animation time: 208.00000000008367
220.00000000002728 - RAF callback, animation time: 209.00000000006003
221.00000000000364 - RAF callback, animation time: 210.00000000003638
221.99999999998 - RAF callback, animation time: 211.00000000001273
222.99999999995634 - RAF callback, animation time: 211.9999999999891
223.9999999999327 - RAF callback, animation time: 212.99999999996544
224.99999999990905 - RAF callback, animation time: 213.9999999999418
225.9999999998854 - RAF callback, animation time: 214.99999999991815
227.00000000008913 - RAF callback, animation time: 216.00000000012187
228.00000000006548 - RAF callback, animation time: 217.00000000009823
229.00000000004184 - RAF callback, animation time: 218.00000000007458
230.0000000000182 - RAF callback, animation time: 219.00000000005093
230.99999999999454 - RAF callback, animation time: 220.00000000002728
231.9999999999709 - RAF callback, animation time: 221.00000000000364
232.99999999994725 - RAF callback, animation time: 221.99999999998
233.9999999999236 - RAF callback, animation time: 222.99999999995634
234.99999999989996 - RAF callback, animation time: 223.9999999999327
236.00000000010368 - RAF callback, animation time: 225.00000000013642
237.00000000008004 - RAF callback, animation time: 226.00000000011278
238.0000000000564 - RAF callback, animation time: 227.00000000008913
239.00000000003274 - RAF callback, animation time: 228.00000000006548
240.0000000000091 - RAF callback, animation time: 229.00000000004184
240.99999999998545 - RAF callback, animation time: 230.0000000000182
241.9999999999618 - RAF callback, animation time: 230.99999999999454
242.99999999993815 - RAF callback, animation time: 231.9999999999709
243.9999999999145 - RAF callback, animation time: 232.99999999994725
244.99999999989086 - RAF callback, animation time: 233.9999999999236
246.0000000000946 - RAF callback, animation time: 235.00000000012733
247.00000000007094 - RAF callback, animation time: 236.00000000010368
248.0000000000473 - RAF callback, animation time: 237.00000000008004
249.00000000002365 - RAF callback, animation time: 238.0000000000564
250 - RAF callback, animation time: 239.00000000003274
250.99999999997635 - RAF callback, animation time: 240.0000000000091
251.9999999999527 - RAF callback, animation time: 240.99999999998545
252.99999999992906 - RAF callback, animation time: 241.9999999999618
253.9999999999054 - RAF callback, animation time: 242.99999999993815
254.99999999988177 - RAF callback, animation time: 243.9999999999145
256.0000000000855 - RAF callback, animation time: 245.00000000011823
257.00000000006185 - RAF callback, animation time: 246.0000000000946
258.0000000000382 - RAF callback, animation time: 247.00000000007094
259.00000000001455 - RAF callback, animation time: 248.0000000000473
259.9999999999909 - RAF callback, animation time: 249.00000000002365
260.99999999996726 - RAF callback, animation time: 250
261.9999999999436 - RAF callback, animation time: 250.99999999997635
262.99999999991996 - RAF callback, animation time: 251.9999999999527
263.9999999998963 - RAF callback, animation time: 252.99999999992906
265.00000000010004 - RAF callback, animation time: 254.0000000001328
266.0000000000764 - RAF callback, animation time: 255.00000000010914
267.00000000005275 - RAF callback, animation time: 256.0000000000855
268.0000000000291 - RAF callback, animation time: 257.00000000006185
269.00000000000546 - RAF callback, animation time: 258.0000000000382
269.9999999999818 - RAF callback, animation time: 259.00000000001455
270.99999999995816 - RAF callback, animation time: 259.9999999999909
271.9999999999345 - RAF callback, animation time: 260.99999999996726
272.99999999991087 - RAF callback, animation time: 261.9999999999436
273.9999999998872 - RAF callback, animation time: 262.99999999991996
275.00000000009095 - RAF callback, animation time: 264.0000000001237
276.0000000000673 - RAF callback, animation time: 265.00000000010004
277.00000000004366 - RAF callback, animation time: 266.0000000000764
278.00000000002 - RAF callback, animation time: 267.00000000005275
278.99999999999636 - RAF callback, animation time: 268.0000000000291
279.9999999999727 - RAF callback, animation time: 269.00000000000546
280.99999999994907 - RAF callback, animation time: 269.9999999999818
281.9999999999254 - RAF callback, animation time: 270.99999999995816
282.9999999999018 - RAF callback, animation time: 271.9999999999345
284.0000000001055 - RAF callback, animation time: 273.00000000013824
285.00000000008185 - RAF callback, animation time: 274.0000000001146
286.0000000000582 - RAF callback, animation time: 275.00000000009095
287.00000000003456 - RAF callback, animation time: 276.0000000000673
288.0000000000109 - RAF callback, animation time: 277.00000000004366
288.99999999998727 - RAF callback, animation time: 278.00000000002
289.9999999999636 - RAF callback, animation time: 278.99999999999636
290.99999999994 - RAF callback, animation time: 279.9999999999727
291.9999999999163 - RAF callback, animation time: 280.99999999994907
292.9999999998927 - RAF callback, animation time: 281.9999999999254
294.0000000000964 - RAF callback, animation time: 283.00000000012915
295.00000000007276 - RAF callback, animation time: 284.0000000001055
296.0000000000491 - RAF callback, animation time: 285.00000000008185
297.00000000002547 - RAF callback, animation time: 286.0000000000582
298.0000000000018 - RAF callback, animation time: 287.00000000003456
298.9999999999782 - RAF callback, animation time: 288.0000000000109
299.9999999999545 - RAF callback, animation time: 288.99999999998727
300.9999999999309 - RAF callback, animation time: 289.9999999999636
301.99999999990723 - RAF callback, animation time: 290.99999999994
302.9999999998836 - RAF callback, animation time: 291.9999999999163
304.0000000000873 - RAF callback, animation time: 293.00000000012005
305.00000000006366 - RAF callback, animation time: 294.0000000000964
306.00000000004 - RAF callback, animation time: 295.00000000007276
307.00000000001637 - RAF callback, animation time: 296.0000000000491
307.9999999999927 - RAF callback, animation time: 297.00000000002547
308.9999999999691 - RAF callback, animation time: 298.0000000000018
309.99999999994543 - RAF callback, animation time: 298.9999999999782
310.9999999999218 - RAF callback, animation time: 299.9999999999545
311.99999999989814 - RAF callback, animation time: 300.9999999999309
313.00000000010186 - RAF callback, animation time: 302.0000000001346
314.0000000000782 - RAF callback, animation time: 303.00000000011096
315.00000000005457 - RAF callback, animation time: 304.0000000000873
316.0000000000309 - RAF callback, animation time: 305.00000000006366
317.0000000000073 - RAF callback, animation time: 306.00000000004
317.99999999998363 - RAF callback, animation time: 307.00000000001637
318.99999999996 - RAF callback, animation time: 307.9999999999927
319.99999999993634 - RAF callback, animation time: 308.9999999999691
320.9999999999127 - RAF callback, animation time: 309.99999999994543
321.99999999988904 - RAF callback, animation time: 310.9999999999218
323.00000000009277 - RAF callback, animation time: 312.0000000001255
324.0000000000691 - RAF callback, animation time: 313.00000000010186
325.0000000000455 - RAF callback, animation time: 314.0000000000782
326.0000000000218 - RAF callback, animation time: 315.00000000005457
326.9999999999982 - RAF callback, animation time: 316.0000000000309
327.99999999997453 - RAF callback, animation time: 317.0000000000073
328.9999999999509 - RAF callback, animation time: 317.99999999998363
329.99999999992724 - RAF callback, animation time: 318.99999999996
330.9999999999036 - RAF callback, animation time: 319.99999999993634
332.0000000001073 - RAF callback, animation time: 321.00000000014006
333.0000000000837 - RAF callback, animation time: 322.0000000001164
334.00000000006 - RAF callback, animation time: 323.00000000009277
335.0000000000364 - RAF callback, animation time: 324.0000000000691
336.00000000001273 - RAF callback, animation time: 325.0000000000455
336.9999999999891 - RAF callback, animation time: 326.0000000000218
337.99999999996544 - RAF callback, animation time: 326.9999999999982
338.9999999999418 - RAF callback, animation time: 327.99999999997453
339.99999999991815 - RAF callback, animation time: 328.9999999999509
340.9999999998945 - RAF callback, animation time: 329.99999999992724
342.0000000000982 - RAF callback, animation time: 331.00000000013097
343.0000000000746 - RAF callback, animation time: 332.0000000001073
345.0000000000273 - RAF callback, animation time: 334.00000000006
347.99999999995634 - RAF callback, animation time: 336.9999999999891
348.9999999999327 - RAF callback, animation time: 337.99999999996544
349.99999999990905 - RAF callback, animation time: 338.9999999999418
350.9999999998854 - RAF callback, animation time: 339.99999999991815
352.00000000008913 - RAF callback, animation time: 341.0000000001219
353.0000000000655 - RAF callback, animation time: 342.0000000000982
354.00000000004184 - RAF callback, animation time: 343.0000000000746
355.0000000000182 - RAF callback, animation time: 344.00000000005093
355.99999999999454 - RAF callback, animation time: 345.0000000000273
356.9999999999709 - RAF callback, animation time: 346.00000000000364
357.99999999994725 - RAF callback, animation time: 346.99999999998
358.9999999999236 - RAF callback, animation time: 347.99999999995634
359.99999999989996 - RAF callback, animation time: 348.9999999999327
361.0000000001037 - RAF callback, animation time: 350.0000000001364
362.00000000008004 - RAF callback, animation time: 351.0000000001128
363.0000000000564 - RAF callback, animation time: 352.00000000008913
368.9999999999145 - RAF callback, animation time: 357.99999999994725
369.99999999989086 - RAF callback, animation time: 358.9999999999236
371.0000000000946 - RAF callback, animation time: 360.00000000012733
372.00000000007094 - RAF callback, animation time: 361.0000000001037
373.0000000000473 - RAF callback, animation time: 362.00000000008004
374.00000000002365 - RAF callback, animation time: 363.0000000000564
375 - RAF callback, animation time: 364.00000000003274
375.99999999997635 - RAF callback, animation time: 365.0000000000091
376.9999999999527 - RAF callback, animation time: 365.99999999998545
377.99999999992906 - RAF callback, animation time: 366.9999999999618
378.9999999999054 - RAF callback, animation time: 367.99999999993815
379.99999999988177 - RAF callback, animation time: 368.9999999999145
381.0000000000855 - RAF callback, animation time: 370.00000000011823
382.00000000006185 - RAF callback, animation time: 371.0000000000946
383.0000000000382 - RAF callback, animation time: 372.00000000007094
384.00000000001455 - RAF callback, animation time: 373.0000000000473
387.99999999991996 - RAF callback, animation time: 376.9999999999527
387.99999999991996 - RAF callback, animation time: 376.9999999999527
388.9999999998963 - RAF callback, animation time: 377.99999999992906
390.00000000010004 - RAF callback, animation time: 379.0000000001328
391.0000000000764 - RAF callback, animation time: 380.00000000010914
392.00000000005275 - RAF callback, animation time: 381.0000000000855
393.0000000000291 - RAF callback, animation time: 382.00000000006185
394.00000000000546 - RAF callback, animation time: 383.0000000000382
394.9999999999818 - RAF callback, animation time: 384.00000000001455
395.99999999995816 - RAF callback, animation time: 384.9999999999909
396.9999999999345 - RAF callback, animation time: 385.99999999996726
397.99999999991087 - RAF callback, animation time: 386.9999999999436
398.9999999998872 - RAF callback, animation time: 387.99999999991996
400.00000000009095 - RAF callback, animation time: 389.0000000001237
401.0000000000673 - RAF callback, animation time: 390.00000000010004
402.00000000004366 - RAF callback, animation time: 391.0000000000764
403.00000000002 - RAF callback, animation time: 392.00000000005275
403.99999999999636 - RAF callback, animation time: 393.0000000000291
404.9999999999727 - RAF callback, animation time: 394.00000000000546
405.99999999994907 - RAF callback, animation time: 394.9999999999818
406.9999999999254 - RAF callback, animation time: 395.99999999995816
407.9999999999018 - RAF callback, animation time: 396.9999999999345
409.0000000001055 - RAF callback, animation time: 398.00000000013824
410.00000000008185 - RAF callback, animation time: 399.0000000001146
411.0000000000582 - RAF callback, animation time: 400.00000000009095
411.0000000000582 - Running checks for time: 400, delay: 9.094947017729282e-11
411.0000000000582 - Checking expectation: [null,0.4,"box","webkitTransform.5",0,0]
412.00000000003456 - Ending test
